# api/openapi.yaml
openapi: 3.1.0
info:
  title: kbdb API
  description: >
    API for tracking a personal mechanical keyboard collection: keyboards,
    switches, keycap sets, and the builds that combine them. Routes are
    nested under /v1/users/{userId}/... to support both a caller's own
    collection and items other users have explicitly shared with them —
    see the userId path parameter's description for the exact rules.
  version: "1.0.0"
  contact:
    url: https://mykeebs.info
  license:
    name: MIT
    url: https://github.com/rogueserenity/kbdb/blob/main/LICENSE
servers:
  - url: https://api.mykeebs.info
security:
  - BearerAuth: []
...
paths:
...
  /v1/users/{userId}/keycap-sets:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [keycap-sets]
      summary: List keycap sets
      operationId: listKeycapSets
      security:
        - {}
        - BearerAuth: []
      description: >
        If userId is the caller's own subject, lists all of the caller's keycap sets. Otherwise, lists only that user's keycap sets explicitly shared with the caller (possibly empty).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of keycap sets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapSetListPage"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [keycap-sets]
      summary: Create a keycap set
      operationId: createKeycapSet
      description: >
        userId must be the caller's own subject; creating in another user's collection always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeycapSetInput"
      responses:
        "201":
          description: The created keycap set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapSet"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
  /v1/users/{userId}/builds:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [builds]
      summary: List builds
      operationId: listBuilds
      security:
        - {}
        - BearerAuth: []
      description: >
        If userId is the caller's own subject, lists all of the caller's builds. Otherwise, lists only that user's builds explicitly shared with the caller (possibly empty).
      parameters:
        - $ref: "#/components/parameters/BuildKeyboardIdFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of builds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildListPage"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [builds]
      summary: Create a build
      operationId: createBuild
      description: >
        userId must be the caller's own subject; creating in another user's collection always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildInput"
      responses:
        "201":
          description: The created build.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/builds/{buildId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/BuildId"
    get:
      tags: [builds]
      summary: Get a build
      operationId: getBuild
      security:
        - {}
        - BearerAuth: []
      description: >
        Returns the build if it belongs to the caller, or if its owner has shared it with the caller. Returns 404 (not 403) if the build does not exist, is not owned by the caller, and is not shared with the caller — this avoids revealing that an unshared item exists.
      responses:
        "200":
          description: The requested build.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    put:
      tags: [builds]
      summary: Update a build
      operationId: updateBuild
      description: >
        userId must be the caller's own subject; updating another user's build always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildInput"
      responses:
        "200":
          description: The updated build.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    delete:
      tags: [builds]
      summary: Delete a build
      operationId: deleteBuild
      description: >
        userId must be the caller's own subject; deleting another user's build always returns 404.
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
  /v1/users/{userId}/builds/{buildId}/images/{imageId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/BuildId"
      - name: imageId
        in: path
        required: true
        schema:
          type: string
        description: The image_id returned when the image was added.
    delete:
      tags: [builds]
      summary: Remove an image from a build
      operationId: deleteBuildImage
      description: >
        userId must be the caller's own subject; removing an image from
        another user's build always returns 404.
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An IdP-issued OIDC JWT (Authorization Bearer <token>).
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
      description: >
        The IdP subject of the item's owner — the `sub` claim of an access
        token, and the `user_id` field of a profile. If this matches the
        authenticated caller's own subject, the request is against their own
        collection. If it names a different user, the request is against
        that user's collection: reads return only items that user has
        explicitly shared with the caller (or, for a specific item, only if
        that item is shared with the caller — otherwise a 404, not a 403,
        to avoid revealing that an item exists but isn't shared); writes
        (create/update/delete) always fail with a 404 in this case, since a
        caller can never modify another user's items. Not a PII/security
        concern to expose in a URL — an IdP subject is an opaque identifier,
        not derived from email or any other identifying attribute.
...
    BuildId:
      name: buildId
      in: path
      required: true
      schema:
        type: string
      description: Server-generated identifier for the build.
...
    BuildKeyboardIdFilter:
      name: keyboard_id
      in: query
      schema:
        type: string
      description: If set, only return builds whose keyboard field is this keyboard's id.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response's next_cursor.
  responses:
    ProblemResponse:
      description: An error, formatted per RFC 9457 (Problem Details for HTTP APIs).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFoundResponse:
      description: The requested resource does not exist, or does not belong to the authenticated user.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ConflictResponse:
      description: >
        The keycap set or build was modified by a concurrent request while
        this one was in flight, and retries were exhausted. Safe to retry
        the request as-is.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details.
      properties:
        type:
          type: string
          format: uri
          example: https://mykeebs.info/errors/validation-error
        title:
          type: string
          example: Validation Error
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        invalid_params:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              reason:
                type: string
              allowed_values:
                type: array
                items:
                  type: string
        blocking_build_ids:
          type: array
          description: >
            Present on a 409 response caused by the item still being
            referenced by one or more builds. Lists the ids of the blocking
            builds.
          items:
            type: string
      required: [type, title, status]
...
    Money:
      type: number
      format: double
      description: A price in the collection's tracking currency.
      minimum: 0
      maximum: 100000000000000
      example: 123.45
...
    Visibility:
      type: string
      enum: [public, authenticated, private]
      description: >
        Who can read this item via its GET routes (create/update/delete
        always require the caller to be the item's owner, regardless of
        visibility). "public": readable with no authentication at all.
        "authenticated": readable by any signed-in user, but not
        anonymous callers. "private": readable only by the owner. There
        is no per-user allowlist/sharing mechanism beyond these three
        tiers — visibility is not scoped to specific other users.
...
    KeycapKit:
      type: object
      required: [kit_id, name]
      properties:
        kit_id:
          type: string
          description: Server-generated identifier for this kit, unique within the keycap set.
        name:
          type: string
        purchase:
          $ref: "#/components/schemas/Purchase"
        image:
          $ref: "#/components/schemas/KeycapKitImage"
          description: >
            Managed via the dedicated image endpoint, not via PUT on
            the kit itself — see POST/DELETE
            .../keycap-sets/{keycapSetId}/kits/{kitId}/image.
    KeycapSetInput:
      type: object
      required: [brand, name, visibility]
      additionalProperties: false
      properties:
        brand:
          type: string
          minLength: 1
          pattern: "\\S"
        name:
          type: string
          minLength: 1
          pattern: "\\S"
        profile:
          type: string
          description: Open vocabulary — validated against the "keycap_profile" lookup at request time.
        material:
          type: string
          description: Open vocabulary — validated against the "keycap_material" lookup at request time.
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
    KeycapSet:
      type: object
      required: [id, brand, name]
      properties:
        id:
          type: string
        brand:
          type: string
        name:
          type: string
        profile:
          type: string
        material:
          type: string
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this keycap set. Only ever present for the
            set's owner - no preference can expose it to another
            caller.
        kits:
          type: array
          items:
            $ref: "#/components/schemas/KeycapKit"
          description: >
            Managed via the dedicated kit endpoints below, not via PUT
            on the set itself — see POST/PUT/DELETE
            .../keycap-sets/{keycapSetId}/kits[/{kitId}].
        primary_kit_id:
          type: [string, "null"]
          description: >
            The kit_id, among this set's kits, whose image represents
            the set in a list view. Null if never set, or if the kit
            it named no longer exists — match against kits[].kit_id
            to find that kit's image.
        order_status:
          type: [string, "null"]
          description: >
            Derived, read-only summary of every kit's
            purchase.order_status: the least-progressed status among
            them wins (Planned < Ordered < Shipped < Delivered), so the
            set isn't shown as Delivered while a kit is still en
            route. A kit with status Cancelled is excluded from that
            comparison unless every kit is Cancelled, in which case
            this is Cancelled too. Null if no kit has an order_status
            set, or the set has no kits.
    KeycapSetSummary:
      type: object
      properties:
        id:
          type: string
        brand:
          type: string
        name:
          type: string
        profile:
          type: string
        order_status:
          type: [string, "null"]
          description: >
            Same derived aggregate as KeycapSet.order_status - see there
            for the bubble-up rule.
        primary_kit_image:
          $ref: "#/components/schemas/KeycapKitImage"
          description: >
            The primary kit's image, if the set has one designated and it
            has an image set. Omitted otherwise — including when the
            designated primary kit no longer exists.
        total_cost:
          $ref: "#/components/schemas/Money"
          description: >
            Total cost of the set's kits with a known purchase price.
            Kits with no known price are excluded rather than treated as
            zero, so this may undercount. Present for the owner only if
            their show_price_to_me preference is set, and for any other
            caller only if the owner's show_price_to_others is.
        currency:
          $ref: "#/components/schemas/Currency"
          description: >
            The owner's display currency, for labelling total_cost.
            Present exactly when total_cost is.
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this keycap set. Set on the keycap set itself,
            not derived from its kits, which have no visibility of their
            own. Only ever present for the set's owner - no preference
            can expose it to another caller.
    KeycapSetListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KeycapSetSummary"
        next_cursor:
          type: [string, "null"]
...
    BuildInput:
      type: object
      required: [keyboard, visibility]
      additionalProperties: false
      properties:
        keyboard:
          type: string
          minLength: 1
          pattern: "\\S"
          description: ID of a Keyboard resource.
        plate:
          type: string
          description: Which of the keyboard's design.plates options is installed.
        case_mount_type:
          $ref: "#/components/schemas/BuildCaseMountType"
        stabs:
          $ref: "#/components/schemas/BuildStabsInput"
        foam:
          type: boolean
        switches:
          type: array
          items:
            $ref: "#/components/schemas/BuildSwitchEntry"
        keycap_kits:
          type: array
          items:
            $ref: "#/components/schemas/BuildKeycapKitEntry"
        build_date:
          type: string
          format: date
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
...
    Build:
      type: object
      description: >
        The read representation of a build. Unlike BuildInput, keyboard,
        switches[].switch, and keycap_kits[].keycap_set are denormalized
        into resolved reference objects rather than bare ids, so a client
        can render a build without follow-up requests.
      required: [id]
      properties:
        id:
          type: string
        keyboard:
          $ref: "#/components/schemas/BuildKeyboardRef"
          description: >
            Omitted if the referenced Keyboard no longer exists (e.g.
            deleted after the build referenced it).
        plate:
          type: string
          description: Which of the keyboard's design.plates options is installed.
        case_mount_type:
          $ref: "#/components/schemas/BuildCaseMountType"
        stabs:
          $ref: "#/components/schemas/BuildStabs"
        foam:
          type: boolean
        switches:
          type: array
          items:
            $ref: "#/components/schemas/BuildSwitchEntryResolved"
        keycap_kits:
          type: array
          items:
            $ref: "#/components/schemas/BuildKeycapKitEntryResolved"
        build_date:
          type: string
          format: date
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this build. Only ever present for the build's
            owner - no preference can expose it to another caller.
        images:
          type: array
          items:
            $ref: "#/components/schemas/BuildImage"
          description: >
            Managed via the dedicated image endpoints below, not via
            PUT on the build itself — see POST/DELETE
            .../builds/{buildId}/images[/{imageId}].
        total_cost:
          $ref: "#/components/schemas/Money"
          description: >
            Total cost of the build's priced components (keyboard,
            switches, keycap kits, stabs). Components with no known price
            are excluded rather than treated as zero, so this may
            undercount.
        currency:
          $ref: "#/components/schemas/Currency"
          description: >
            The owner's display currency, for labelling total_cost.
            Present exactly when total_cost is.
...
    BuildListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BuildSummary"
        next_cursor:
          type: [string, "null"]
...